Added static line chart (work with data will be later)
