+++ package.json
@@ -54,6 +54,7 @@
     ]
   },
   "devDependencies": {
+    "@types/chart.js": "^2.9.37",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
     "prettier": "^2.7.1",
