+++ src/pages/Coin/Coin.tsx
@@ -6,7 +6,7 @@ import { useLocalStore } from "@utils/useLocalStore";
 import { observer } from "mobx-react-lite";
 import { useParams } from "react-router-dom";
 
-import Chart from "./components/Chart/Chart";
+import ChartLine from "./components/Chart/ChartLine";
 import Header from "./components/Header/Header";
 import styles from "./styles.module.scss";
 
@@ -52,7 +52,7 @@ const Coin: React.FC = () => {
           </div>
         </div>
       </div>
-      <Chart />
+      <ChartLine />
       <Card
         image={coinStore.coinData.image}
         title={coinStore.coinData.name}
